evals/pyproject.toml
[project]
name = "skill-eval"
version = "0.1.0"
description = "A/B testing tool for LLM skills"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.scripts]
skill-eval = "skill_eval.cli:app"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
